+++ Fluent.Ribbon/Controls/Button.cs
@@ -264,9 +264,11 @@ namespace Fluent
         /// <summary>
         /// Handles key tip pressed
         /// </summary>
-        public void OnKeyTipPressed()
+        public KeyTipPressedResult OnKeyTipPressed()
         {
             this.OnClick();
+
+            return KeyTipPressedResult.Empty;
         }
 
         /// <summary>