+++ Fluent.Ribbon/Controls/InRibbonGallery.cs
@@ -982,9 +982,19 @@ namespace Fluent
         /// <summary>
         /// Handles key tip pressed
         /// </summary>
-        public void OnKeyTipPressed()
+        public KeyTipPressedResult OnKeyTipPressed()
         {
             this.IsDropDownOpen = true;
+
+            if (this.DropDownPopup?.Child != null)
+            {
+                Keyboard.Focus(this.DropDownPopup.Child);
+                this.DropDownPopup.Child.MoveFocus(new TraversalRequest(FocusNavigationDirection.First));
+
+                return new KeyTipPressedResult(false, true);
+            }
+
+            return new KeyTipPressedResult(false, true);
         }
 
         /// <summary>