+++ Fluent.Ribbon/IKeyTipedControl.cs
@@ -13,7 +13,7 @@
         /// <summary>
         /// Handles key tip pressed
         /// </summary>
-        void OnKeyTipPressed();
+        KeyTipPressedResult OnKeyTipPressed();
 
         /// <summary>
         /// Handles back navigation with KeyTips